requirements.txt
langchain==1.2.0
langchain_community==0.4.1
langchain_core==1.2.4
langchain_google_genai==4.1.2
Pillow==12.0.0
python-dotenv==1.2.1
python_docx==1.2.0
streamlit==1.50.0
